Rework tests for better consumer visibility.
This patch moves as many tests as viable to tests/, in order to make
nuking "non-critical" code easier.

( Nothing will hurt if you nuke tests/*, just you won't have as many
  tests )

All tests that require src/normalization_tests.rs are gated by

```rust
  #[cfg(bloaty_test)]
```

So that they only run when invoked as:

  RUSTFLAGS="--cfg bloaty_test" cargo test

If you do this, its your reponsibility to make sure that file is there.

This means nuking src/normalization_tests.rs doesn't break what happens
by default when a consumer runs `cargo test` in a published crate.

One tests that previously lived in its own file was folded back
into src/lib.rs, because it was more concise that way, and after the
migration to tests/,  it was pretty much there on its own, not
warranting an individual file.